// src/hooks/orval/model/participantsRecord.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * PocketBase API
 * API documentation generated from PocketBase schema with collection grouping
 * OpenAPI spec version: 1.0.0
 */
import type { ParticipantsRecordAddressPhase } from './participantsRecordAddressPhase.ts';
import type { ParticipantsRecordPerformanceBasis } from './participantsRecordPerformanceBasis.ts';
import type { ParticipantsRecordPreferredLane } from './participantsRecordPreferredLane.ts';
import type { ParticipantsRecordPreferredRolesItem } from './participantsRecordPreferredRolesItem.ts';
import type { ParticipantsRecordRegistrationStatus } from './participantsRecordRegistrationStatus.ts';
import type { ParticipantsRecordRoleRankings } from './participantsRecordRoleRankings.ts';
import type { ParticipantsRecordStatus } from './participantsRecordStatus.ts';
import type { ParticipantsRecordTeamIntent } from './participantsRecordTeamIntent.ts';

export interface ParticipantsRecord {
  /** Unique record identifier */
  id?: string;
  /** Relation field. When expanded, this becomes the related record object. Use the 'expand' query parameter to expand relations. */
  tournament: string;
  name: string;
  email: string;
  ign: string;
  birthdate: string;
  contact_number?: string;
  user_id: string;
  server_id: string;
  address_phase: ParticipantsRecordAddressPhase;
  address_package: string;
  address_block: string;
  address_lot: string;
  preferred_lane: ParticipantsRecordPreferredLane;
  /** @maxItems 5 */
  preferred_roles?: ParticipantsRecordPreferredRolesItem[];
  role_rankings?: ParticipantsRecordRoleRankings;
  performance_basis?: ParticipantsRecordPerformanceBasis;
  team_intent?: ParticipantsRecordTeamIntent;
  /** Relation field. When expanded, this becomes the related record object. Use the 'expand' query parameter to expand relations. */
  preferred_team?: string;
  preferred_team_name?: string;
  registration_status: ParticipantsRecordRegistrationStatus;
  registration_reject_reason?: string;
  registration_status_code?: string;
  consent_version?: string;
  school_id_front?: string;
  school_id_back?: string;
  purok_endorsement?: string;
  status: ParticipantsRecordStatus;
  /** Relation field. When expanded, this becomes the related record object. Use the 'expand' query parameter to expand relations. */
  team?: string;
  /** Relation field. When expanded, this becomes the related record object. Use the 'expand' query parameter to expand relations. */
  created_by?: string;
  /** Relation field. When expanded, this becomes the related record object. Use the 'expand' query parameter to expand relations. */
  updated_by?: string;
  archived: boolean;
  consent_accepted_at?: string;
  created?: string;
  updated?: string;
}