// src/hooks/orval/model/teamsRecordStatus.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * PocketBase API
 * API documentation generated from PocketBase schema with collection grouping
 * OpenAPI spec version: 1.0.0
 */

export type TeamsRecordStatus = typeof TeamsRecordStatus[keyof typeof TeamsRecordStatus];


export const TeamsRecordStatus = {
  forming: 'forming',
  ready: 'ready',
  incomplete: 'incomplete',
  inactive: 'inactive',
} as const;